Even more user-interface things calculated using floats
Yes, this work is snow-balling, but in a good direction:)

- More and more UI controls calculate things using floats (which makes them
  working nicer when UI scaling is used).
- And some more properties are exposed using floats (which makes them
  more useful when UI scaling is used, and there's no point in
  limiting stuff like padding or margin to being integers, since they
  will be multiplied by float UIScale anyway).

We want to eventually calculate everything in UI using floats, to
support UI scaling without any glitches.

This commit *does* cause some compatibility break in public CGE API,
by changing some properties from Integer/Cardinal to Single. User code
depending in some way that e.g. Padding is Integer (doing something
like "MyInteger := Button.Padding") will need to be adjusted. This is
a conscious breakage -- creating a float-alternative for *every* property would
complicate our implementation too much, so we do this only for some
crucial things (like Left, Bottom, Width, Height, ScreenRect...), the
rest is just changed to use Single.

+++ src/fonts/opengl/castlefonts.pas
@@ -190,9 +190,9 @@ type
       (from the line FirstToBreak).
 
       @groupBegin }
-    procedure BreakLines(const unbroken: string; broken: TStrings; MaxLineWidth: integer); overload;
-    procedure BreakLines(unbroken, broken: TStrings; MaxLineWidth: integer); overload;
-    procedure BreakLines(broken: TStrings; MaxLineWidth: Integer; FirstToBreak: integer); overload;
+    procedure BreakLines(const unbroken: string; broken: TStrings; MaxLineWidth: Single); overload;
+    procedure BreakLines(unbroken, broken: TStrings; MaxLineWidth: Single); overload;
+    procedure BreakLines(broken: TStrings; MaxLineWidth: Single; FirstToBreak: integer); overload;
     { @groupEnd }
 
     { Largest width of the line of text in given list.
@@ -668,7 +668,7 @@ begin
 end;
 
 procedure TCastleFont.BreakLines(const unbroken: string;
-  broken: TStrings; MaxLineWidth: integer);
+  broken: TStrings; MaxLineWidth: Single);
 var
   unbrokenlist: TStringList;
 begin
@@ -680,7 +680,7 @@ begin
 end;
 
 procedure TCastleFont.BreakLines(unbroken, broken: TStrings;
-  MaxLineWidth: integer);
+  MaxLineWidth: Single);
 var
   i, FirstToBreak: Integer;
 begin
@@ -690,7 +690,7 @@ begin
 end;
 
 procedure TCastleFont.BreakLines(broken: TStrings;
-  MaxLineWidth: Integer; FirstToBreak: integer);
+  MaxLineWidth: Single; FirstToBreak: integer);
 var
   I: Integer;
   LineWidth, LineWidthBytes: Integer;
